requirements.txt
fastapi
azure-cosmos
uvicorn